Split build graph lowering DSL
