+++ docs/PHASE_PLAN.md
@@ -2380,6 +2380,12 @@ checked-in docs, tests, and commits only.
   `src/typechecker/expressions/function_checking.rs`, keeping
   `src/typechecker/expressions.rs` focused on expression dispatch while
   preserving return/fallthrough and defer coverage.
+- Build graph lowering DSL spellings and target-field ownership now live in
+  `src/build_graph/lowering/dsl.rs`, reducing
+  `src/build_graph/lowering.rs` below the 500-line cleanup threshold while
+  preserving the existing build.zen lowering behavior. The guard
+  `production_rust_files_stay_below_cleanup_threshold` keeps the focused
+  production files from silently growing back past the cleanup threshold.
 
 ## Current Phase
