+++ src/build_graph/lowering.rs
@@ -1,139 +1,9 @@
-use std::str::FromStr;
-
 use crate::ast::{Declaration, Expression, MatchArm, Program, Statement};
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum BuildTargetDslKind {
-    Executable,
-    Test,
-    Library,
-}
+#[path = "lowering/dsl.rs"]
+mod dsl;
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum BuildTargetField {
-    Name,
-    Main,
-    Root,
-    RootSourceFile,
-    OutDir,
-    Dependencies,
-    Features,
-    Exports,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum BuildTargetDslIdent {
-    Builder,
-    Add,
-    Build,
-    Env,
-    Os,
-    ReadFile,
-}
-
-impl BuildTargetDslKind {
-    const ALL: [Self; 3] = [Self::Executable, Self::Test, Self::Library];
-    const EXECUTABLE: &'static str = "Executable";
-    const TEST: &'static str = "Test";
-    const LIBRARY: &'static str = "Library";
-
-    fn as_str(self) -> &'static str {
-        match self {
-            Self::Executable => Self::EXECUTABLE,
-            Self::Test => Self::TEST,
-            Self::Library => Self::LIBRARY,
-        }
-    }
-
-    fn supported_display_list() -> String {
-        let names = Self::ALL
-            .iter()
-            .map(|kind| format!("`{kind}`"))
-            .collect::<Vec<_>>();
-        let Some((last, rest)) = names.split_last() else {
-            return String::new();
-        };
-        if rest.is_empty() {
-            return last.clone();
-        }
-        format!("{}, and {last}", rest.join(", "))
-    }
-}
-
-impl BuildTargetField {
-    const NAME: &'static str = "name";
-    const MAIN: &'static str = "main";
-    const ROOT: &'static str = "root";
-    const ROOT_SOURCE_FILE: &'static str = "root_source_file";
-    const OUT_DIR: &'static str = "out_dir";
-    const DEPENDENCIES: &'static str = "dependencies";
-    const FEATURES: &'static str = "features";
-    const EXPORTS: &'static str = "exports";
-
-    fn as_str(self) -> &'static str {
-        match self {
-            Self::Name => Self::NAME,
-            Self::Main => Self::MAIN,
-            Self::Root => Self::ROOT,
-            Self::RootSourceFile => Self::ROOT_SOURCE_FILE,
-            Self::OutDir => Self::OUT_DIR,
-            Self::Dependencies => Self::DEPENDENCIES,
-            Self::Features => Self::FEATURES,
-            Self::Exports => Self::EXPORTS,
-        }
-    }
-}
-
-impl fmt::Display for BuildTargetField {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.write_str(self.as_str())
-    }
-}
-
-impl BuildTargetDslIdent {
-    const BUILDER: &'static str = "b";
-    const ADD: &'static str = "add";
-    const BUILD: &'static str = "build";
-    const ENV: &'static str = "env";
-    const OS: &'static str = "os";
-    const READ_FILE: &'static str = "read_file";
-
-    fn as_str(self) -> &'static str {
-        match self {
-            Self::Builder => Self::BUILDER,
-            Self::Add => Self::ADD,
-            Self::Build => Self::BUILD,
-            Self::Env => Self::ENV,
-            Self::Os => Self::OS,
-            Self::ReadFile => Self::READ_FILE,
-        }
-    }
-}
-
-impl fmt::Display for BuildTargetDslIdent {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.write_str(self.as_str())
-    }
-}
-
-impl fmt::Display for BuildTargetDslKind {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.write_str(self.as_str())
-    }
-}
-
-impl FromStr for BuildTargetDslKind {
-    type Err = ();
-
-    fn from_str(value: &str) -> Result<Self, Self::Err> {
-        match value {
-            Self::EXECUTABLE => Ok(Self::Executable),
-            Self::TEST => Ok(Self::Test),
-            Self::LIBRARY => Ok(Self::Library),
-            _ => Err(()),
-        }
-    }
-}
+use dsl::{BuildTargetDslIdent, BuildTargetDslKind, BuildTargetField};
 
 impl BuildGraph {
     pub fn from_build_program(program: &Program) -> Result<Self, BuildGraphError> {
